+++ src/main/java/humbledude/lotto/domain/LottoNumberSet.java
@@ -1,6 +1,5 @@
 package humbledude.lotto.domain;
 
-import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
